chore: move auth.ts to utils folder, update build
