+++ backend/tsconfig.json
@@ -50,6 +50,7 @@
           "services/*",
           "interfaces/*",
           "views/*",
+          "utils/*",
         ]
     },
     // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
